refactor(cli): rename render-table to render-solidity
