+++ packages/cli/src/index.ts
@@ -1,8 +1,8 @@
 export { parseStoreConfig } from "./config/parseStoreConfig.js";
 export { loadStoreConfig } from "./config/loadStoreConfig.js";
 export { loadWorldConfig, resolveWorldConfig, parseWorldConfig } from "./config/loadWorldConfig.js";
-export { renderTablesFromConfig } from "./render-table/renderTablesFromConfig.js";
-export { renderTable } from "./render-table/renderTable.js";
+export { renderTablesFromConfig } from "./render-solidity/renderTablesFromConfig.js";
+export { renderTable } from "./render-solidity/renderTable.js";
 
 export type {
   StoreUserConfig,
