+++ packages/cli/src/render-solidity/renderTable.ts
